requirements.txt
aioserial==1.3.1
influxdb-client==1.48.0
python-dotenv==1.1.0